Fix: Logger deadlock with recursive_mutex and TensorRT 10.1 / CUDA 13.2 configuration

+++ src/core/main.cpp
@@ -185,7 +185,7 @@ int main(int argc, char* argv[]) {
                             "). Idle senders remain available to keep channel names stable in vMix.");
         }
         
-        const int channelsToInit = std::min(kMaxSpoutChannels, static_cast<int>(numDevices));
+        const int channelsToInit = (std::min)(static_cast<int>(kMaxSpoutChannels), static_cast<int>(numDevices));
         for (int i = 0; i < channelsToInit; i++) {
             auto capture = std::make_unique<DeckLinkCapture>();
             const std::string channelName = "Channel_" + std::to_string(i + 1);
